pyproject.toml
[project]
name = "ques"
version = "0.1.0"
description = ""
authors = [
    {name = "Dastan",email = "dastan.kubanuulu1@gmail.com"}
]
readme = "README.md"
package-mode = false
requires-python = ">=3.13"
dependencies = [
    "django (>=5.2,<6.0)",
    "pillow (>=11.1.0,<12.0.0)",
    "djangorestframework (>=3.16.0,<4.0.0)",
    "django-cors-headers (>=4.7.0,<5.0.0)",
    "django-filter (>=25.1,<26.0)",
    "django-redis (>=5.4.0,<6.0.0)",
    "djangorestframework-simplejwt (>=5.5.0,<6.0.0)",
    "drf-spectacular (>=0.28.0,<0.29.0)",
    "gunicorn (>=23.0.0,<24.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "python-decouple (>=3.8,<4.0)",
    "requests (>=2.32.3,<3.0.0)",
    "requests (>=2.32.3,<3.0.0)",
    "sqlalchemy (>=2.0.40,<3.0.0)",
    "whitenoise (>=6.9.0,<7.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
